@@ -15,24 +15,25 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 3.2.0
+version: 4.0.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  path_provider: ^1.6.24
-  provider: ^4.3.2+2
-  shared_preferences: ^0.5.12+4
-  share: ^0.6.5+4
-  url_launcher: ^5.7.10
-  webfeed: ^0.6.0
-  jiffy: ^3.0.1
-  sqflite: ^1.3.0
-  animate_do: ^2.0.0
+  http:
+  path_provider:
+  provider:
+  shared_preferences:
+  share:
+  url_launcher:
+  webfeed:
+  jiffy:
+  sqflite:
+  animate_do:
+  flutter_lints:
 
 
   # The following adds the Cupertino Icons font to your application.
