@@ -3,6 +3,7 @@ import 'package:rssyoutubefschmtz/util/changelog.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class AppInfoPage extends StatelessWidget {
+  const AppInfoPage({Key? key}) : super(key: key);
 
   _launchGithub()  {
     const url = 'https://github.com/Fschmatz/rssyoutubefschmtz';
@@ -12,11 +13,11 @@ class AppInfoPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
 
-    Color themeColorApp = Theme.of(context).accentColor;
+    Color themeColorApp = Theme.of(context).colorScheme.primary;
 
     return Scaffold(
         appBar: AppBar(
-          title: Text("App Info"),
+          title: const Text("App Info"),
           elevation: 0,
         ),
         body: ListView(children: <Widget>[
@@ -24,7 +25,7 @@ class AppInfoPage extends StatelessWidget {
           CircleAvatar(
             radius: 55,
             backgroundColor: Colors.green[400],
-            child: CircleAvatar(
+            child: const CircleAvatar(
               radius: 50,
               backgroundImage: AssetImage('assets/avatar.jpg'),
             ),
@@ -40,7 +41,7 @@ class AppInfoPage extends StatelessWidget {
           const SizedBox(height: 15),
           const Divider(),
           ListTile(
-            leading: SizedBox(
+            leading: const SizedBox(
               height: 0.1,
             ),
             title: Text("Dev".toUpperCase(),
@@ -49,7 +50,7 @@ class AppInfoPage extends StatelessWidget {
                     fontWeight: FontWeight.w700,
                     color: themeColorApp)),
           ),
-          ListTile(
+          const ListTile(
             leading: Icon( Icons.info_outline),
             title: Text(
               "HAMMERED AND REDONE: 0.5 Times !!!",
@@ -58,7 +59,7 @@ class AppInfoPage extends StatelessWidget {
               ),
             ),
           ),
-          ListTile(
+          const ListTile(
             leading: SizedBox(
               height: 0.1,
             ),
@@ -71,7 +72,7 @@ class AppInfoPage extends StatelessWidget {
           ),
           const Divider(),
           ListTile(
-            leading: SizedBox(
+            leading: const SizedBox(
               height: 0.1,
             ),
             title: Text("Source Code".toUpperCase(),
@@ -82,15 +83,15 @@ class AppInfoPage extends StatelessWidget {
           ),
           ListTile(
             onTap: () {_launchGithub();},
-            leading: Icon(Icons.open_in_new_outlined),
-            title: Text("View on GitHub",
+            leading: const Icon(Icons.open_in_new_outlined),
+            title: const Text("View on GitHub",
                 style: TextStyle(
                     decoration: TextDecoration.underline,
                     color: Colors.blue)),
           ),
           const Divider(),
           ListTile(
-            leading: SizedBox(
+            leading: const SizedBox(
               height: 0.1,
             ),
             title: Text("Quote".toUpperCase(),
@@ -99,7 +100,7 @@ class AppInfoPage extends StatelessWidget {
                     fontWeight: FontWeight.w700,
                     color: themeColorApp)),
           ),
-          ListTile(
+          const ListTile(
             leading: Icon(Icons.messenger_outline),
             title: Text(
               "No one in the brief history of computing has ever written a piece of perfect software. It’s unlikely that you’ll be the first.\nAndy Hunt ",
