@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:rssyoutubefschmtz/util/changelog.dart';
 import 'package:rssyoutubefschmtz/util/theme.dart';
-import 'appInfoPage.dart';
-import 'changelogPage.dart';
+import 'app_info_page.dart';
+import 'changelog_page.dart';
 
 class SettingsPage extends StatefulWidget {
   @override
   _SettingsPageState createState() => _SettingsPageState();
 
-  SettingsPage({Key key}) : super(key: key);
+  const SettingsPage({Key? key}) : super(key: key);
 }
 
 class _SettingsPageState extends State<SettingsPage> {
@@ -21,9 +21,13 @@ class _SettingsPageState extends State<SettingsPage> {
   @override
   Widget build(BuildContext context) {
 
-    Color themeColorApp = Theme.of(context).accentColor;
+    Color themeColorApp = Theme.of(context).colorScheme.primary;
 
     return Scaffold(
+        appBar: AppBar(
+          title: const Text("Settings"),
+          elevation: 0,
+        ),
         body: SingleChildScrollView(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
@@ -32,20 +36,20 @@ class _SettingsPageState extends State<SettingsPage> {
                 elevation: 1,
                 margin: const EdgeInsets.fromLTRB(16, 20, 16, 25),
                 color: themeColorApp,
-                shape: RoundedRectangleBorder(
+                shape: const RoundedRectangleBorder(
                   borderRadius: BorderRadius.all(Radius.circular(15)),
                 ),
                 child: ListTile(
                   title: Text(
                     Changelog.appName + " " + Changelog.appVersion,
                     textAlign: TextAlign.center,
-                    style: TextStyle(fontSize: 17.5, color: Colors.black),
+                    style: const TextStyle(fontSize: 17.5, color: Colors.black),
                   ),
                 ),
               ),
               const Divider(),
               ListTile(
-                leading: SizedBox(height: 0.1,),
+                leading: const SizedBox(height: 0.1,),
                 title:    Text(
                     "About".toUpperCase(),
                     style: TextStyle(
@@ -55,10 +59,10 @@ class _SettingsPageState extends State<SettingsPage> {
                 ),
               ),
               ListTile(
-                leading: Icon(
+                leading: const Icon(
                   Icons.info_outline,
                 ),
-                title: Text(
+                title: const Text(
                   "App Info",
                   style: TextStyle(fontSize: 16),
                 ),
@@ -66,7 +70,7 @@ class _SettingsPageState extends State<SettingsPage> {
                   Navigator.push(
                       context,
                       MaterialPageRoute<void>(
-                        builder: (BuildContext context) => AppInfoPage(),
+                        builder: (BuildContext context) => const AppInfoPage(),
                         fullscreenDialog: true,
                       ));
                 },
@@ -75,10 +79,10 @@ class _SettingsPageState extends State<SettingsPage> {
                 height: 10.0,
               ),
               ListTile(
-                leading: Icon(
+                leading: const Icon(
                   Icons.text_snippet_outlined,
                 ),
-                title: Text(
+                title: const Text(
                   "Changelog",
                   style: TextStyle(fontSize: 16),
                 ),
@@ -86,14 +90,14 @@ class _SettingsPageState extends State<SettingsPage> {
                   Navigator.push(
                       context,
                       MaterialPageRoute<void>(
-                        builder: (BuildContext context) => ChangelogPage(),
+                        builder: (BuildContext context) => const ChangelogPage(),
                         fullscreenDialog: true,
                       ));
                 },
               ),
               const Divider(),
               ListTile(
-                leading: SizedBox(height: 0.1,),
+                leading: const SizedBox(height: 0.1,),
                 title:    Text(
                     "General".toUpperCase(),
                     style: TextStyle(
@@ -104,11 +108,11 @@ class _SettingsPageState extends State<SettingsPage> {
               ),
               Consumer<ThemeNotifier>(
                 builder: (context, notifier, child) => SwitchListTile(
-                    title: Text(
+                    title: const Text(
                       "Dark Theme",
                       style: TextStyle(fontSize: 16),
                     ),
-                    secondary: Icon(Icons.brightness_6_outlined),
+                    secondary: const Icon(Icons.brightness_6_outlined),
                     activeColor: Colors.blue,
                     value: notifier.darkTheme,
                     onChanged: (value) {
