@@ -3,5 +3,5 @@ class ChannelFeed{
   final String name;
   final String linkFeed;
 
-  ChannelFeed({this.name,this.linkFeed});
+  ChannelFeed({required this.name,required this.linkFeed});
 }
